+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.3.0(unreleased)
+- remove support for bing,bard,pandora
+- refactor chat logic, add self host model support
+- support new image generation endpoint
+- admin system to manage langchain(flowise backend)
+
 ## 1.2.0
 - rename `api_key` to `openai_api_key` in `config.json`
 - rename `bing_api_endpoint` to `api_endpoint` in `config.json` and `env` file